Rename and reorganize mock modules
