+++ lib/davinci_dtr_test_kit/dtr_smart_app_suite.rb
@@ -5,7 +5,7 @@ require_relative 'auth_groups/oauth2_authentication_group'
 require_relative 'client_groups/resp_assist_device/dtr_smart_app_questionnaire_workflow_group'
 require_relative 'client_groups/dinner_static/dtr_smart_app_questionnaire_workflow_group'
 require_relative 'client_groups/dinner_adaptive/dtr_smart_app_questionnaire_workflow_group'
-require_relative 'endpoints/mock_authorization/handlers'
+require_relative 'endpoints/mock_authorization/mock_authorization'
 require_relative 'endpoints/mock_authorization/authorize_endpoint'
 require_relative 'endpoints/mock_authorization/token_endpoint'
 require_relative 'endpoints/mock_payer/questionnaire_package_endpoint'
