+++ lib/davinci_dtr_test_kit/endpoints/mock_authorization/mock_authorization.rb
@@ -1,9 +1,21 @@
-require_relative '../../urls'
-require_relative 'common'
-
 module DaVinciDTRTestKit
   module MockAuthorization
+    RSA_PRIVATE_KEY = OpenSSL::PKey::RSA.generate(2048)
+    RSA_PUBLIC_KEY = RSA_PRIVATE_KEY.public_key
+    SUPPORTED_SCOPES = ['launch', 'patient/*.rs', 'user/*.rs', 'offline_access', 'openid', 'fhirUser'].freeze
+
     module_function
+
+    def extract_client_id_from_bearer_token(request)
+      token = request.headers['authorization']&.delete_prefix('Bearer ')
+      jwt =
+        begin
+          JWT.decode(token, nil, false)
+        rescue StandardError
+          nil
+        end
+      jwt&.first&.dig('inferno_client_id')
+    end
 
     def auth_server_jwks(_env)
       response_body = {
@@ -59,8 +71,6 @@ module DaVinciDTRTestKit
       }.to_json
       [200, { 'Content-Type' => 'application/json', 'Access-Control-Allow-Origin' => '*' }, [response_body]]
     end
-
-    private
 
     def env_base_url(env, endpoint_path)
       protocol = env['rack.url_scheme']
